Added UI layer
Added a class UserInterface to handle userInput.

Also refactored the creation of Objects (Booking, Service and Barber) to be handled by the Barbershop class adhering to the creator design pattern.

Updated Dummy populator to create a barbershop object populated with barbers and services.

Also made Barber and Consumer inherit the User class